--- src/moe/utils/gating.py
@@ -1,2 +1,0 @@
-"""Gating functions (softmax routing, top-k selection, hard routing, 
-temperature scaling) are shared across Soft, Sparse, and Hard MoE heads"""
